lint: replace inspect_err
inspect_err is unstable in 1.76.0 Rust toolchain

+++ attestation-agent/attester/src/tsm_report/mod.rs
@@ -75,9 +75,9 @@ impl TsmReportPath {
         // path (rmdir way) when TsmReportPath instance goes out of scope.
         let path = p.into_path();
 
-        check_tsm_report_provider(path.as_path(), wanted).inspect_err(|_| {
+        if check_tsm_report_provider(path.as_path(), wanted).is_err() {
             let _ = std::fs::remove_dir(path.as_path());
-        })?;
+        }
 
         Ok(Self { path })
     }
